+++ src/ElementTransformer.php
@@ -2,20 +2,28 @@
 
 namespace craft\elementapi;
 
+use craft\base\Element;
 use craft\base\ElementInterface;
-use craft\helpers\ArrayHelper;
 use League\Fractal\TransformerAbstract;
 
 /**
  * Element transformer class.
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since  2.0
+ * @since 2.0
  */
 class ElementTransformer extends TransformerAbstract
 {
     public function transform(ElementInterface $element): array
     {
-        return ArrayHelper::toArray($element);
+        // Get the serialized custom field values
+        $fields = $element->getSerializedFieldValues();
+
+        // Get the element attributes that aren't custom fields
+        /** @var Element $element */
+        $attributes = array_diff($element->attributes(), array_keys($fields));
+
+        // Return the element as an array merged with its serialized custom field values
+        return array_merge($element->toArray($attributes), $fields);
     }
 }